Run pre-training on AWS Sagemaker
Add code to send DistilBERT model for pre-training]
on AWS Sagemaker, requirements.txt is now put into
a subfolder which indicates which training task it is for.
Separated the run_mlm.py from the run_mlm_aws.py
since there are differences in the batch size and other
arguments for the AWS training.