Clean up train and test set generation
Fix bugs in functions for generating train
and test sets, namely the mixing of SSOC
descriptions in the test set, the lack of data
type check/enforcement for the SSOC column,
and add in a new functionality to enable
the function to work on non-augmented data.
Also fix formatting / spelling errors.