Feature pretrain language model (#21)
* Start notebook

Create notebook for testing pretraining of
DistilBERT on MCF data

* Add masked language modelling script

Adapt MLM script from huggingface to be used
on the local computer. Add arguments for training
and evaluating progress for pre-training. Include
code for generating small subset of pre-training data.

* Small changes to pre-training notebook

Fix title of notebook and amend code to output
full pretraining text file.

* Run pre-training on AWS Sagemaker

Add code to send DistilBERT model for pre-training]
on AWS Sagemaker, requirements.txt is now put into
a subfolder which indicates which training task it is for.
Separated the run_mlm.py from the run_mlm_aws.py
since there are differences in the batch size and other
arguments for the AWS training.

* Clean up Jupyter Notebook

Cleaned up the AWS pre-training section

Co-authored-by: shaunkhoo <shaun@dsaid.gov.sg>