Modify function for generating predictions
Streamlined core code for generating prediction from
the model into 3 functions - one for a single prediction,
one for multiple predictions, and one which reads in the
model, tokenizer, encoding, and generates a single prediction
based on input text (this is for the Lambda function).
Documentation still needs to be done.